+++ projects/igniteui-angular/src/lib/grids/grid/grid.component.ts
@@ -656,6 +656,7 @@ export class IgxGridComponent extends IgxGridBaseDirective implements GridType, 
      * this.grid.groupRowTemplate = myRowTemplate;
      * ```
      */
+    @Input()
     public get groupRowTemplate(): TemplateRef<IgxGroupByRowTemplateContext> {
         return this._groupRowTemplate;
     }