+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes for each version of this project will be documented in this 
 ### New Features
 - `igxGrid` - exposing new Input properties:
 
+    - `excelStyleHeaderIconTemplate` - Gets/Sets the excel style header icon.
+    - `groupRowTemplate` - Gets/Sets the template reference for the group row.
     - `headSelectorTemplate` - Gets/Sets the header row selector template.
     - `rowSelectorTemplate` - Gets/Sets the custom template used for row selectors.
     - `groupByRowSelectorTemplate` - Gets/Sets the custom template used for the group row selectors.